Extract some models
