+++ src/audioset/Audioset.ts
@@ -1,21 +1,11 @@
+import { AudiosetBundle } from "./AudiosetBundle";
+import { AudiosetVisuals } from "./AudiosetVisuals";
 import { Clip } from "./Clip";
 import { Track } from "./Track";
 
-/**
- * An unsepecified audioset type
- */
-export interface AudiosetBundle {
-  format: "atpls-audioset";
-  type: "audioset" | "project";
-  meta: AudiosetMetadata;
-}
-
-export function isAudiosetBundle(data: any): data is AudiosetBundle {
-  return typeof data === "object" && data.format === "atpls-audioset";
-}
-
-export interface Audioset extends AudiosetBundle {
+export interface Audioset {
   type: "audioset";
+  meta: AudiosetMetadata;
   id: string;
   tracks: Track[];
   clips: Clip[];
@@ -23,7 +13,7 @@ export interface Audioset extends AudiosetBundle {
   index: AudiosetIndexes;
 }
 
-export function isAudiosetPack(audioset: AudiosetBundle): audioset is Audioset {
+export function isAudioset(audioset: AudiosetBundle): audioset is Audioset {
   return audioset.type === "audioset";
 }
 
@@ -38,49 +28,26 @@ export interface AudiosetMetadata {
   bpm?: number;
 }
 /**
- * The aaudioset metadata extended to include the audioset id and publish path
+ * The audioset metadata extended to include the audioset id and publish path
  * It's used in audioset project to list audiosets
  */
-export type AudiosetReference = AudiosetMetadata & {
+export interface AudiosetReference extends AudiosetMetadata {
   id: string;
   publish_path: string;
-};
+}
 
 /**
  * An AudiosetProject is a group of audiosets
  */
-export interface AudiosetProject extends AudiosetBundle {
+export interface AudiosetProject {
   type: "project";
+  meta: AudiosetMetadata;
   audiosets: AudiosetReference[];
 }
+
 export interface AudiosetIndexes {
   clipById: Record<string, Clip>;
   trackById: Record<string, Track>;
   clipIdsOfTrack: Record<string, string[]>;
   trackIdOfClip: Record<string, string>;
-}
-
-export type AudiosetVisuals = MapVisuals | PanelVisuals;
-
-export interface MapVisuals {
-  mode: "map";
-  geomap: {
-    url: string;
-    scaleFactor: number;
-    center: {
-      x: number;
-      y: number;
-    };
-  };
-}
-
-export interface PanelVisuals {
-  mode: "panel";
-  image: {
-    url: string;
-    size: {
-      width: number;
-      height: number;
-    };
-  };
 }
