+++ src/audioset/index.ts
@@ -1,5 +1,7 @@
 export * from "./Clip";
 export * from "./Track";
+export * from "./AudiosetBundle";
 export * from "./Audioset";
+export * from "./AudiosetVisuals";
 export * from "./EmptyAudioset";
 export * from "./createAudioset";
